Dev to main

+++ src/components/EditFolderPopupContent/index.js
@@ -7,8 +7,11 @@ import { html } from 'htm/react'
 import { MenuItem, MenuList } from './styles'
 import { ConfirmationModalContent } from '../../containers/Modal/ConfirmationModalContent'
 import { CreateFolderModalContent } from '../../containers/Modal/CreateFolderModalContent'
+import { CreateFolderModalContentV2 } from '../../containers/Modal/CreateFolderModalContentV2/CreateFolderModalContentV2'
+import { DeleteFolderModalContentV2 } from '../../containers/Modal/DeleteFolderModalContentV2/DeleteFolderModalContentV2'
 import { useModal } from '../../context/ModalContext'
 import { DeleteIcon, FolderIcon } from '../../lib-react-components'
+import { isV2 } from '../../utils/designVersion'
 
 /**
  *
@@ -19,7 +22,7 @@ import { DeleteIcon, FolderIcon } from '../../lib-react-components'
  */
 export const EditFolderPopupContent = ({ name }) => {
   const { i18n } = useLingui()
-  const { deleteFolder } = useFolders()
+  const { deleteFolder, data: folderData } = useFolders()
   const { setModal, closeModal } = useModal()
 
   const menuItems = useMemo(
@@ -28,34 +31,59 @@ export const EditFolderPopupContent = ({ name }) => {
         name: i18n._('Delete'),
         type: 'delete',
         icon: DeleteIcon,
-        onClick: () =>
-          setModal(
-            html`<${ConfirmationModalContent}
-              primaryAction=${() => {
-                deleteFolder(name)
-                closeModal()
-              }}
-              secondaryAction=${closeModal}
-              title=${i18n._('Are you sure you want to delete this folder?')}
-              text=${i18n._(
-                'This action will permanently delete the folder and all items contained within it. Are you sure you want to proceed?'
-              )}
-            />`
-          )
+        onClick: () => {
+          if (isV2()) {
+            const count =
+              folderData?.customFolders?.[name]?.records?.length ?? 0
+            if (count === 1) {
+              deleteFolder(name)
+              closeModal()
+            } else {
+              setModal(
+                <DeleteFolderModalContentV2
+                  folderName={name}
+                  count={count - 1}
+                  onClose={closeModal}
+                />
+              )
+            }
+          } else {
+            setModal(
+              html`<${ConfirmationModalContent}
+                primaryAction=${() => {
+                  deleteFolder(name)
+                  closeModal()
+                }}
+                secondaryAction=${closeModal}
+                title=${i18n._('Are you sure you want to delete this folder?')}
+                text=${i18n._(
+                  'This action will permanently delete the folder and all items contained within it. Are you sure you want to proceed?'
+                )}
+              />`
+            )
+          }
+        }
       },
       {
         name: i18n._('Rename'),
         type: 'rename',
         icon: FolderIcon,
         onClick: () =>
-          setModal(
-            html`<${CreateFolderModalContent}
-              initialValues=${{ title: name }}
-            />`
-          )
+          isV2()
+            ? setModal(
+                <CreateFolderModalContentV2
+                  initialValues={{ title: name }}
+                  onClose={closeModal}
+                />
+              )
+            : setModal(
+                html`<${CreateFolderModalContent}
+                  initialValues=${{ title: name }}
+                />`
+              )
       }
     ],
-    [closeModal, deleteFolder, i18n, name, setModal]
+    [closeModal, deleteFolder, folderData, i18n, name, setModal]
   )
 
   const handleMenuItemClick = (e, item) => {
